Use milliseconds for prop date values in database

+++ common/db/entity/PropSchema.js
@@ -21,7 +21,7 @@ const PropSchema = new EntitySchema({
       type: 'int',
     },
     date: {
-      type: 'int',
+      type: 'bigint', // milliseconds
     },
     x: {
       type: 'double precision', // meters
